:truck: Rename: Checkbox 구현을 CheckSquare 컴포넌트로 이동 (#1)
